changed to floats because floats good

+++ src/reccalc/parse.h
@@ -51,7 +51,7 @@ extern int yydebug;
 		// Whether to print the intermediate results.
 		int verbose;
 		// Value of the last computation.
-		int value;
+		float value;
 		// Number of errors.
 		int nerrs;
 	} result;
@@ -84,9 +84,9 @@ union YYSTYPE
   /* "string"  */
   char* STR;
   /* "number"  */
-  int NUM;
+  float NUM;
   /* exp  */
-  int exp;
+  float exp;
 #line 91 "parse.h"
 
 };
@@ -101,12 +101,12 @@ int yyparse (yyscan_t scanner, result *res);
 /* "%code provides" blocks.  */
 #line 18 "parse.y"
 
-  // Tell Flex the expected prototype of yylex.
-  // The scanner argument must be named yyscanner.
-#define YY_DECL enum yytokentype yylex (YYSTYPE* yylval, yyscan_t yyscanner, result *res)
-  YY_DECL;
+	// Tell Flex the expected prototype of yylex.
+	// The scanner argument must be named yyscanner.
+	#define YY_DECL enum yytokentype yylex (YYSTYPE* yylval, yyscan_t yyscanner, result *res)
+	YY_DECL;
 
-  void yyerror (yyscan_t scanner, result *res, const char *msg, ...);
+	void yyerror (yyscan_t scanner, result *res, const char *msg, ...);
 
 #line 112 "parse.h"
